Python/requirements.txt
# ============================================================================
# CORE MCP DEPENDENCIES
# ============================================================================
mcp>=1.2.0
fastmcp>=0.4.1
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ============================================================================
# ASYNC & PERFORMANCE
# ============================================================================
aiohttp>=3.9.0
# uvloop is NOT supported on Windows, using default asyncio
orjson>=3.9.0
websockets>=12.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
PyJWT>=2.8.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ============================================================================
# CACHING
# ============================================================================
redis>=5.0.0
aiocache>=0.12.0

# ============================================================================
# CODE EXECUTION & SANDBOXING
# ============================================================================
docker>=7.0.0
ipybox>=0.1.0

# ============================================================================
# PROCEDURAL GENERATION
# ============================================================================
noise>=1.2.2
numpy>=1.26.0
scipy>=1.11.0
networkx>=3.2

# ============================================================================
# MACHINE LEARNING & AI
# ============================================================================
torch>=2.1.0
torchvision>=0.16.0
# tensorflow>=2.15.0  # Optional: Uncomment if using TensorFlow
scikit-learn>=1.3.0


# ============================================================================
# VISUALIZATION
# ============================================================================
matplotlib>=3.8.0
graphviz>=0.20.0
# pygraphviz can be difficult to install on Windows without C++ build tools
# pygraphviz>=1.11

# ============================================================================
# MONITORING & OBSERVABILITY
# ============================================================================
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation>=0.43b0

# ============================================================================
# LOGGING & UTILITIES
# ============================================================================
python-dotenv>=1.0.0
loguru>=0.7.0
colorama>=0.4.6
tqdm>=4.66.0

# ============================================================================
# HTTP & API
# ============================================================================
requests>=2.31.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# ============================================================================
# DATA VALIDATION & SERIALIZATION
# ============================================================================
marshmallow>=3.20.0
jsonschema>=4.20.0

# ============================================================================
# TESTING
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
